make sure the global dispatcher is used for Node.js fetch
